requirements.txt
python-dotenv
streamlit
watchdog